src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/set-attribute-in-shadow-crash-expected.txt

CONSOLE ERROR: line 8: Uncaught TypeError: host.createShadowRoot is not a function
Tests that setting an attribute in Shadow DOM does not cause a crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
